Fix minor issues from PR feedback
- Remove unnecessary magic '!' check in Node.Validate
- Remove JetBrains namespace import and fix namespace in SerializableDictionary
- Ensure class name matches filename for SerializableDictionary
- Clarify usage of CacheToAttribute and node caching logic
